# doc/technical/openapi.yaml
openapi: 3.1.0
info:
  title: 'SoManAgent API'
  version: 1.0.0
  description: 'Versioned OpenAPI contract for the SoManAgent backend. This specification is spec-first: operations may be documented ahead of implementation when they explicitly set `x-somanagent-implemented` to `false`.'
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  '/api/projects/{projectId}/chat/{agentId}':
    get:
      tags:
        - Chat
      summary: 'Get chat history with an agent'
      operationId: getChatHistory
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/AgentIdPath'
      responses:
        200:
          $ref: '#/components/responses/Ok'
    post:
      tags:
        - Chat
      summary: 'Send a chat message to an agent'
      operationId: sendChatMessage
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/AgentIdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/projects/{projectId}/chat/{agentId}/reply':
    post:
      tags:
        - Chat
      summary: 'Reply in an agent chat thread'
      operationId: replyChatMessage
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/AgentIdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
  '/api/projects/{projectId}/chat/{agentId}/messages/{messageId}':
    patch:
      tags:
        - Chat
      summary: 'Update a human chat message'
      operationId: updateChatMessage
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/AgentIdPath'
        - $ref: '#/components/parameters/MessageIdPath'
      requestBody:
        $ref: '#/components/requestBodies/ArbitraryJsonBody'
      responses:
        200:
          $ref: '#/components/responses/Ok'
...
components:
  parameters:
...
    AgentIdPath:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MessageIdPath:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
    ArbitraryJsonBody:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArbitraryJsonObject'
...
  responses:
    Ok:
      description: 'Successful response'
...
  schemas:
...
    ArbitraryJsonObject:
      type: object
      additionalProperties: true
...